Game update
A different class for each block type was created.

Now the ball can move around the playing area.

Now the ball can detect collisions.

Now when the ball collides with a block, the block is destroyed.

The classes for ball, paddle and blocks changed.

The scene for the game was updated.